Refactor code for improved performance and readability
